+++ lib/src/models/yust_user.g.dart
@@ -39,7 +39,11 @@ YustUser _$YustUserFromJson(Map json) => YustUser(
       ..authenticationMethod = $enumDecodeNullable(
           _$YustAuthenticationMethodEnumMap, json['authenticationMethod'],
           unknownValue: JsonKey.nullForUndefinedEnumValue)
-      ..domain = json['domain'] as String?;
+      ..domain = json['domain'] as String?
+      ..profilePicture = json['profilePicture'] == null
+          ? null
+          : YustFile.fromJson(
+              Map<String, dynamic>.from(json['profilePicture'] as Map));
 
 Map<String, dynamic> _$YustUserToJson(YustUser instance) => <String, dynamic>{
       'id': instance.id,
@@ -61,6 +65,7 @@ Map<String, dynamic> _$YustUserToJson(YustUser instance) => <String, dynamic>{
           _$YustAuthenticationMethodEnumMap[instance.authenticationMethod],
       'domain': instance.domain,
       'authId': instance.authId,
+      'profilePicture': instance.profilePicture?.toJson(),
     };
 
 const _$YustGenderEnumMap = {