+++ lib/src/util/google_cloud_helpers.dart
@@ -1,0 +1,2 @@
+export 'google_cloud_helpers_dart.dart'
+    if (dart.library.ui) './google_cloud_helpers_flutter.dart';